fix: merge with same name entity
