+++ src/main/java/com/sparta/trello/domain/column/entity/TrelloColumn.java
@@ -1,33 +1,39 @@
 package com.sparta.trello.domain.column.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import com.sparta.trello.domain.card.entity.Card;
+import com.sparta.trello.domain.common.entity.BaseEntity;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Entity
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class TrelloColumn {
+public class TrelloColumn extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	private String title;
 
-	private int position;
+	private Integer position;
 
-	// @ManyToOne
-	// @JoinColumn(name = "board_id")
-	// private Board board;
+	// card와 매핑
+	@OneToMany(mappedBy = "trelloColumn", cascade = CascadeType.ALL)
+	private List<Card> cards;
 
+	public TrelloColumn updatePosition(Integer newPosition) {
+		return TrelloColumn.builder()
+				.id(this.id)
+				.title(this.title)
+				.position(newPosition)
+				.build();
+	}
 }
